+++ gmd/Cui/AddStashDlg.cs
@@ -5,17 +5,17 @@ namespace gmd.Cui;
 
 interface IAddStashDlg
 {
-    R<string> Show(string message);
+    R<string> Show();
 }
 
 
 class AddStashDlg : IAddStashDlg
 {
-    public R<string> Show(string message)
+    public R<string> Show()
     {
         var dlg = new UIDialog("Stash", 44, 7);
         dlg.AddLabel(1, 0, "Stash Message:");
-        var name = dlg.AddTextField(1, 1, 40, $"on '{message}'");
+        var name = dlg.AddTextField(1, 1, 40, "");
 
         if (!dlg.ShowOkCancel(name)) return R.Error();